+++ R/RcppExports.R
@@ -711,6 +711,12 @@ copyDatasetFiles <- function(new_filename, old_filename, format = "") {
     .Call(`_gdalraster__getCreationOptions`, format)
 }
 
+#' @noRd
+NULL
+
+#' @noRd
+NULL
+
 #' Is GEOS available?
 #'
 #' `has_geos()` returns a logical value indicating whether GDAL was built