+++ NEWS.md
@@ -1,5 +1,11 @@
 # gdalraster 1.5.1 (dev)
 
+* add `buildRAT()`: build a new GDAL Raster Attribute Table as data frame (2023-10-08)
+
+* add `GDALRaster$setDefaultRAT()`: set Raster Attribute Table from data frame (2023-10-08)
+
+* gdalraster now requires GDAL >= 2.4.0 (previously >= 2.3.0) (2023-10-08)
+
 * add `GDALRaster$getDefaultRAT()`: return Raster Attribute Table as a data frame (2023-10-04)
 
 * add `GDALRaster$getDefaultHistogram()`: fetch default raster histogram (2023-10-02)